[base] Simplify autohinting conditions.
After FreeType switched type1 and t1cid drivers to the Adobe engine
a while ago, we could flag them as DRIVER_HINTS_LIGHTLY.  We do it now
explicitly in the driver_init and the property_set functions  as a
future reminder. This simplifies the conditions for autohinting.

* src/base/ftobjs.c (FT_Load_Glyph): Rely only on !DRIVER_HINTS_LIGHTLY
when selecting the autohinter for the LIGHT mode.
* src/cff/cffdrivr.c (cff_driver_class): Move the flag from here...
* src/cff/cffobjs.c (cff_driver_init): ... to here, explicitly.
* src/cid/cidobjs.c (sid_driver_init): Set it explicitly.
* src/type1/t1objs.c (T1_Driver_Init): Set it explicitly.
* src/base/ftpsprop.c (ps_property_set): Sync the flags when switching.